+++ scripts/make/file-content/componentRef.ts
@@ -1,5 +1,5 @@
-import { pascalToCamel } from '../utils/stringUtils';
 import { STYLE_PREFIX } from '../constants';
+import { pascalToCamel } from '../utils/stringUtils';
 
 /* prettier-ignore */
 export const componentRef = (componentName: string) =>
@@ -18,7 +18,7 @@ export const ${componentName} = forwardRef<HTMLButtonElement, ${componentName}Pr
   ({ someProp = 'Default value', className, ...buttonProps }, ref) => {
     const ${pascalToCamel(componentName)}Class = classnames(
       '${STYLE_PREFIX}__base ${STYLE_PREFIX}__${pascalToCamel(componentName)}',
-      { ['${STYLE_PREFIX}__${pascalToCamel(componentName)}--disabled']: buttonProps.disabled },
+      { '${STYLE_PREFIX}__${pascalToCamel(componentName)}--disabled': buttonProps.disabled },
       className,
     );