+++ src/components/DragLayer/DragLayer.stories.tsx
@@ -7,11 +7,14 @@ import docs from './README.md';
 
 export default { title: 'DragLayer', parameters: { info: { text: docs, propTablesExclude: [Draggable] } } };
 
+const WIDTH = 200;
+const HEIGHT = 50;
+
 const commonStyle: CSSProperties = {
   padding: 16,
   background: 'lightgray',
-  width: 200,
-  height: 50,
+  width: WIDTH,
+  height: HEIGHT,
   textAlign: 'center',
 };
 
@@ -21,6 +24,23 @@ export const basic = () => (
       <div style={{ ...commonStyle, border: '1px solid red' }}>This div is draggable!</div>
     </Draggable>
     <DragLayer>
+      <div style={{ ...commonStyle, border: '1px solid blue', opacity: 0.9 }}>Custom preview!</div>
+    </DragLayer>
+  </DndProvider>
+);
+
+export const withCustomTranslate = () => (
+  <DndProvider backend={Backend}>
+    <Draggable index={0} hideDragPreview={true}>
+      <div style={{ ...commonStyle, border: '1px solid red' }}>This div is draggable!</div>
+    </Draggable>
+    <DragLayer
+      customTranslate={({ mousePosition }) => {
+        const x = mousePosition.x - WIDTH / 2;
+        const y = mousePosition.y - HEIGHT / 2;
+        return { x, y };
+      }}
+    >
       <div style={{ ...commonStyle, border: '1px solid blue', opacity: 0.9 }}>Custom preview!</div>
     </DragLayer>
   </DndProvider>