+++ src/components/DragLayer/DragLayer.tsx
@@ -1,6 +1,6 @@
 import classnames from 'classnames';
-import React, { CSSProperties, FC, HTMLAttributes, useMemo, ReactNode } from 'react';
-import { useDragLayer } from 'react-dnd';
+import React, { CSSProperties, FC, HTMLAttributes, ReactNode, useMemo } from 'react';
+import { useDragLayer, XYCoord } from 'react-dnd';
 
 export interface DragLayerProps extends HTMLAttributes<HTMLDivElement> {
   /**
@@ -9,19 +9,27 @@ export interface DragLayerProps extends HTMLAttributes<HTMLDivElement> {
    * will see both.
    */
   children?: ReactNode;
+  /**
+   * If not given, will use `currentOffset.x` and `currentOffset.y`. Can return
+   * custom translate coordinates. This allows you to always center on mouse
+   * position, or translate to any coordinates.
+   */
+  customTranslate?: (params: { currentOffset: XYCoord; mousePosition: XYCoord }) => { x: number; y: number };
 }
 
-export const DragLayer: FC<DragLayerProps> = ({ children, className, ...divProps }) => {
-  const { isDragging, currentOffset } = useDragLayer(monitor => ({
+export const DragLayer: FC<DragLayerProps> = ({ children, customTranslate, className, ...divProps }) => {
+  const { currentOffset, isDragging, mousePosition } = useDragLayer(monitor => ({
     currentOffset: monitor.getSourceClientOffset(),
     isDragging: monitor.isDragging(),
+    mousePosition: monitor.getClientOffset(),
   }));
 
   const style = useMemo<CSSProperties>(() => {
-    if (!currentOffset) return { display: 'none' };
-    const transform = `translate(${currentOffset.x}px, ${currentOffset.y}px)`;
+    if (!currentOffset || !mousePosition) return { display: 'none' };
+    const { x, y } = customTranslate?.({ currentOffset, mousePosition }) ?? currentOffset;
+    const transform = `translate(${x}px, ${y}px)`;
     return { transform, WebkitTransform: transform };
-  }, [currentOffset]);
+  }, [currentOffset, customTranslate, mousePosition]);
 
   if (!isDragging) return null;